Move Bookmarks window to its own class, add fullscreen bookmarks
Users can now pick which view they prefer in the settings window.
