+++ ContentEditor.App/Imgui/Tools/PakBrowser.cs
@@ -48,25 +48,13 @@ public partial class PakBrowser(ContentWorkspace contentWorkspace, string[]? pak
     private ListFileWrapper? currentList;
     private ListFileWrapper? activeListFile;
     private BookmarkHolder _bookmarks = new BookmarkHolder(contentWorkspace);
-    private List<string> _activeTagFilter = new();
-    private string bookmarkSearch = string.Empty;
-    private bool isBookmarkSearchMatchCase = false;
-    private string customBookmarkComment = "";
-    private string? editingCustomBookmark = null;
+    private BookmarksPanel? _bookmarksPanel;
+    private BookmarksPanel BookmarksPanel => _bookmarksPanel ??= new BookmarksPanel(contentWorkspace, _bookmarks, path => CurrentDir = path);
+    private bool isShowBookmarks = false;
     private bool jumpToPageTop = false;
-    private enum FilterMode
-    {
-        AnyMatch,
-        AllMatch,
-        ExactMatch
-    }
-    private FilterMode _filterMode = FilterMode.AnyMatch;
-
     private PakReader? unpacker;
     private int unpackExpectedFiles;
-
     private bool hasInvalidatedPaks;
-    private bool isShowBookmarks = false;
     private int itemsPerPage = 1000;
 
     private ImGuiSortDirection gridSortDir = ImGuiSortDirection.Ascending;
@@ -260,10 +248,13 @@ public partial class PakBrowser(ContentWorkspace contentWorkspace, string[]? pak
         bool isHideCustoms = _bookmarks.User.IsHideBookmarks;
         bool isBookmarked = _bookmarks.User.IsBookmarked(Workspace.Config.Game.name, CurrentDir);
         ImguiHelpers.AlignElementRight((ImGui.CalcTextSize($"{AppIcons.SI_ViewGridSmall}").X + ImGui.GetStyle().FramePadding.X * 2) * 2 + ImGui.GetStyle().ItemSpacing.X);
-        ImguiHelpers.ToggleButton($"{AppIcons.SI_Bookmarks}", ref isShowBookmarks, Colors.IconActive);
+        bool isBookmarksActive = AppConfig.Instance.UseBookmarkWindow ? EditorWindow.CurrentWindow?.HasSubwindow<BookmarksWindow>(out _) == true : isShowBookmarks;
+        if (ImguiHelpers.ToggleButton($"{AppIcons.SI_Bookmarks}", ref isBookmarksActive, Colors.IconActive)) {
+            ToggleBookmarks();
+        }
         ImguiHelpers.Tooltip("Bookmarks"u8);
         if (ImGui.IsWindowFocused(ImGuiFocusedFlags.RootAndChildWindows) && AppConfig.Instance.Key_PakBrowser_OpenBookmarks.Get().IsPressed()) {
-            isShowBookmarks = !isShowBookmarks;
+            ToggleBookmarks();
         }
         ImGui.SameLine();
         if (ImGui.Button(DisplayMode == FileDisplayMode.Grid ? $"{AppIcons.SI_ViewGridSmall}" : $"{AppIcons.List}")) {
@@ -273,99 +264,8 @@ public partial class PakBrowser(ContentWorkspace contentWorkspace, string[]? pak
         ImguiHelpers.Tooltip(DisplayMode == FileDisplayMode.Grid ? "Grid View"u8 : "List View"u8);
         ImGui.Spacing();
         ImGui.Separator();
-        if (isShowBookmarks) {
-            ImguiHelpers.ToggleButtonMultiColor(AppIcons.SIC_BookmarkHide, ref isHideDefaults, [Colors.IconSecondary, Colors.IconPrimary, Colors.IconPrimary], Colors.IconActive);
-            ImguiHelpers.Tooltip("Hide Default Bookmarks"u8);
-            _bookmarks.Defaults.IsHideBookmarks = isHideDefaults;
-            using (var _ = ImguiHelpers.Disabled(_bookmarks.User.GetBookmarks(Workspace.Config.Game.name).Count == 0)) {
-                ImGui.SameLine();
-                ImguiHelpers.ToggleButtonMultiColor(AppIcons.SIC_BookmarkCustomHide, ref isHideCustoms, [Colors.IconSecondary, Colors.IconPrimary, Colors.IconPrimary], Colors.IconActive);
-                ImguiHelpers.Tooltip("Hide Custom Bookmarks"u8);
-                _bookmarks.User.IsHideBookmarks = isHideCustoms;
-                ImGui.SameLine();
-                if (ImguiHelpers.ButtonMultiColor(AppIcons.SIC_BookmarkCustomClear, [Colors.IconPrimary, Colors.IconTertiary])) {
-                    ImGui.OpenPopup("Confirm Action"u8);
-                }
-                ImguiHelpers.Tooltip("Clear Custom Bookmarks");
-                AppImguiHelpers.ShowActionModal(Lang.General.ConfirmTitle, $"{AppIcons.SI_GenericDelete2}", Colors.IconTertiary,
-                    Lang.PakBrowser.ConfirmDeleteBookmarks.FormatRef(Lang.TranslateGame(Workspace.Config.Game.name)),
-                    () => {
-                        _bookmarks.User.ClearBookmarks(Workspace.Config.Game.name);
-                        Logger.Info($"Cleared custom bookmarks for {Workspace.Config.Game.name}");
-                    }
-                );
-            }
-            ImGui.SameLine();
-            string filterModeName = _filterMode switch {
-                FilterMode.AnyMatch => "Any",
-                FilterMode.AllMatch => "All",
-                FilterMode.ExactMatch => "Exact",
-                _ => "?"
-            };
-            string filterLabelDisplayText = _activeTagFilter.Count == 0 ? $"{AppIcons.SI_Filter}" : $"{AppIcons.SI_Filter} : " + _activeTagFilter.Count.ToString();
-            Vector2 filterLabelSize = ImGui.CalcTextSize(filterLabelDisplayText);
-            float filterComboWidth = filterLabelSize.X + ImGui.GetStyle().FramePadding.X * 2 + ImGui.GetStyle().ItemSpacing.X + ImGui.GetFontSize();
-            float searchBarWidth = 260f;
-            ImguiHelpers.AlignElementRight(((ImGui.CalcTextSize($"{AppIcons.SI_GenericClose}").X + ImGui.GetStyle().FramePadding.X * 2) + ImGui.GetStyle().ItemSpacing.X) * 2 + filterComboWidth + searchBarWidth + ImGui.GetStyle().ItemSpacing.X);
-            ImguiHelpers.ToggleButton($"{AppIcons.SI_GenericMatchCase}", ref isBookmarkSearchMatchCase, Colors.IconActive);
-            ImguiHelpers.Tooltip("Match Case");
-            ImGui.SameLine();
-            ImGui.SetNextItemWidth(searchBarWidth);
-            AppImguiHelpers.ClearableInputText("##BookmarkSearch"u8, $"{AppIcons.SI_GenericMagnifyingGlass} Search Comments", ref bookmarkSearch, 64);
-            ImGui.SameLine();
-            ImGui.SetNextItemWidth(filterComboWidth);
-            if (ImGui.BeginCombo("##TagFilterCombo"u8, filterLabelDisplayText, ImGuiComboFlags.HeightLargest)) {
-                ImGui.TextDisabled("Filter Mode:");
-                ImGui.SameLine();
-
-                if (ImGui.SmallButton(filterModeName)) {
-                    _filterMode = _filterMode switch {
-                        FilterMode.AnyMatch => FilterMode.AllMatch,
-                        FilterMode.AllMatch => FilterMode.ExactMatch,
-                        FilterMode.ExactMatch => FilterMode.AnyMatch,
-                        _ => FilterMode.AnyMatch
-                    };
-                }
-
-                if (ImGui.BeginItemTooltip()) {
-                    ImGui.SeparatorText("Filter Modes");
-                    ImGui.BulletText("Any: Keep entries with at least one matching tag");
-                    ImGui.BulletText("All: Keep entries containing all active tags");
-                    ImGui.BulletText("Exact: Keep entries with tags exactly matching the active filters");
-                    ImGui.EndTooltip();
-                }
-                ImGui.Separator();
-
-                foreach (var tag in BookmarkManager.TagInfoMap.Keys) {
-                    bool isSelected = _activeTagFilter.Contains(tag);
-                    if (ImGui.Checkbox(tag, ref isSelected)) {
-                        if (isSelected) {
-                            _activeTagFilter.Add(tag);
-                        } else {
-                            _activeTagFilter.Remove(tag);
-                        }
-                    }
-                }
-                ImGui.EndCombo();
-            }
-            ImguiHelpers.Tooltip("Filters");
-            ImGui.SameLine();
-            using (var _ = ImguiHelpers.Disabled(_activeTagFilter.Count == 0)) {
-                if (ImguiHelpers.ButtonMultiColor(AppIcons.SIC_FilterClear, [Colors.IconTertiary, Colors.IconPrimary])) {
-                    _activeTagFilter.Clear();
-                }
-                ImguiHelpers.Tooltip("Clear Filters");
-            }
-            if (_bookmarks.Defaults.GetBookmarks(Workspace.Config.Game.name).Count > 0 && !_bookmarks.Defaults.IsHideBookmarks) {
-                ShowBookmarksTable("Default", 3, _bookmarks.Defaults, _activeTagFilter, bookmarkSearch.Trim());
-            }
-            if (_bookmarks.User.GetBookmarks(Workspace.Config.Game.name).Count > 0 && !_bookmarks.User.IsHideBookmarks) {
-                ShowBookmarksTable("Custom", 4, _bookmarks.User, _activeTagFilter, bookmarkSearch.Trim());
-            } else {
-                if (!_bookmarks.User.IsHideBookmarks) {
-                    ImGui.TextDisabled("No Custom Bookmarks yet...");
-                }
-            }
+        if (isShowBookmarks && !AppConfig.Instance.UseBookmarkWindow) {
+            BookmarksPanel.ShowBookmarks();
             ImGui.Separator();
             ImGui.Spacing();
         }
@@ -957,174 +857,6 @@ public partial class PakBrowser(ContentWorkspace contentWorkspace, string[]? pak
             ImGui.EndPopup();
         }
     }
-
-    private void ShowBookmarksTable(string label, int columnNum, BookmarkManager manager, List<string> activeTagFilter, string searchText)
-    {
-        var bookmarks = manager.GetBookmarks(Workspace.Config.Game.name);
-        if (bookmarks.Count == 0) {
-            return;
-        }
-
-        var useCompactFilePaths = AppConfig.Instance.UsePakCompactFilePaths.Get();
-        var filteredBookmarks = FilterBookmarks(bookmarks.ToList(), activeTagFilter, searchText, _filterMode);
-
-        if (filteredBookmarks.Count == 0) {
-            ImGui.SeparatorText(label + " [No Matches Found]");
-            return;
-        }
-
-        float rowHeight = ImGui.GetTextLineHeightWithSpacing();
-        float currDisplayHeight = filteredBookmarks.Count <= 5 ? 0 : rowHeight * 10;
-        var bookmarkTableFlags = ImGuiTableFlags.Resizable | ImGuiTableFlags.BordersInnerV | ImGuiTableFlags.BordersOuterV | ImGuiTableFlags.RowBg;
-        if (filteredBookmarks.Count >= 5) {
-            bookmarkTableFlags |= ImGuiTableFlags.ScrollY;
-        }
-        ImGui.SeparatorText(label);
-        if (ImGui.BeginChild($"{label}_Scroll", new Vector2(0, currDisplayHeight), ImGuiChildFlags.AlwaysAutoResize | ImGuiChildFlags.AutoResizeY)) {
-            if (ImGui.BeginTable($"{label}Table", columnNum, bookmarkTableFlags)) {
-                ImGui.TableSetupColumn("Path"u8, ImGuiTableColumnFlags.WidthStretch);
-                ImGui.TableSetupColumn("Tags"u8, ImGuiTableColumnFlags.WidthStretch);
-                ImGui.TableSetupColumn("Comment"u8, ImGuiTableColumnFlags.WidthStretch);
-                if (manager == _bookmarks.User) {
-                    ImGui.TableSetupColumn("Order"u8, ImGuiTableColumnFlags.WidthFixed, ((ImGui.GetFrameHeight() * 2f) + ImGui.GetStyle().ItemSpacing.X * 3f));
-                }
-                ImGui.TableHeadersRow();
-
-                foreach (var bm in filteredBookmarks) {
-                    string displayPath = useCompactFilePaths ? LessCompactFilePath(bm.Path) : bm.Path;
-                    ImGui.TableNextRow();
-                    ImGui.TableSetColumnIndex(0);
-                    if (ImGui.Selectable(displayPath, false)) {
-                        CurrentDir = bm.Path;
-                    }
-
-                    if (manager == _bookmarks.User && ImGui.BeginPopupContextItem(bm.Path)) {
-                        ShowBookmarksContextMenu(manager, bm);
-                        ImGui.EndPopup();
-                    }
-
-                    ImGui.TableSetColumnIndex(1);
-                    foreach (var tag in bm.Tags) {
-                        ImGui.PushID($"{bm.Path}_{tag}");
-                        if (BookmarkManager.TagInfoMap.TryGetValue(tag, out var info)) {
-                            var cols = info.Colors();
-                            ImGui.PushStyleColor(ImGuiCol.Button, cols[0]);
-                            ImGui.PushStyleColor(ImGuiCol.ButtonHovered, cols[1]);
-                            ImGui.PushStyleColor(ImGuiCol.ButtonActive, cols[2]);
-                            ImGui.PushStyleColor(ImGuiCol.Text, Vector4.One);
-                            if (ImGui.Button(info.Icon)) {
-                                if (!activeTagFilter.Remove(tag)) {
-                                    activeTagFilter.Add(tag);
-                                }
-                            }
-                            ImguiHelpers.Tooltip(tag);
-                            ImGui.PopStyleColor(4);
-                        }
-                        ImGui.SameLine();
-                        ImGui.PopID();
-                    }
-
-                    ImGui.TableSetColumnIndex(2);
-                    if (!string.IsNullOrEmpty(bm.Comment)) {
-                        (manager == _bookmarks.Defaults ? (Action<string>)ImGui.TextDisabled : ImGui.Text)(bm.Comment);
-                    }
-
-                    if (manager == _bookmarks.User) {
-                        ImGui.TableSetColumnIndex(3);
-                        int idx = bookmarks.IndexOf(bm);
-                        int uniqueId = bm.GetHashCode();
-                        int? moveFrom = null, moveTo = null;
-                        using (var _ = ImguiHelpers.Disabled(!string.IsNullOrEmpty(searchText))) {
-                            if (ImGui.ArrowButton($"##up_{uniqueId}", ImGuiDir.Up) && idx > 0) {
-                                moveFrom = idx;
-                                moveTo = idx - 1;
-                            }
-                            ImGui.SameLine();
-                            if (ImGui.ArrowButton($"##down_{uniqueId}", ImGuiDir.Down) && idx < bookmarks.Count - 1) {
-                                moveFrom = idx;
-                                moveTo = idx + 1;
-                            }
-                        }
-
-                        if (moveFrom.HasValue && moveTo.HasValue) {
-                            (bookmarks[moveFrom.Value], bookmarks[moveTo.Value]) = (bookmarks[moveTo.Value], bookmarks[moveFrom.Value]);
-                            manager.SaveBookmarks();
-                        }
-                    }
-                }
-                ImGui.EndTable();
-            }
-        }
-        ImGui.EndChild();
-    }
-    private List<BookmarkManager.BookmarkEntry> FilterBookmarks(List<BookmarkManager.BookmarkEntry> bookmarks, List<string> activeTagFilter, string searchText, FilterMode filterMode)
-    {
-        return bookmarks.Where(bm => {
-            bool tagMatch = true;
-            if (activeTagFilter.Count > 0) {
-                switch (filterMode) {
-                    case FilterMode.AnyMatch:
-                        tagMatch = activeTagFilter.Any(tag => bm.Tags.Contains(tag));
-                        break;
-                    case FilterMode.AllMatch:
-                        tagMatch = activeTagFilter.All(tag => bm.Tags.Contains(tag));
-                        break;
-                    case FilterMode.ExactMatch:
-                        tagMatch = bm.Tags.Count == activeTagFilter.Count && activeTagFilter.All(tag => bm.Tags.Contains(tag));
-                        break;
-                }
-            }
-
-            bool commentMatch = string.IsNullOrEmpty(searchText) || (bm.Comment?.Contains(searchText, isBookmarkSearchMatchCase ? StringComparison.Ordinal : StringComparison.OrdinalIgnoreCase) ?? false);
-            return tagMatch && commentMatch;
-        }).ToList();
-    }
-
-    private void ShowBookmarksContextMenu(BookmarkManager manager, BookmarkManager.BookmarkEntry bm)
-    {
-
-        if (ImguiHelpers.ContextMenuItem("##CopyPath", AppIcons.SI_FileCopyPath, "Copy Path", Colors.IconPrimary)) {
-            EditorWindow.CurrentWindow?.CopyToClipboard(bm.Path);
-        }
-        if (ImguiHelpers.ContextMenuItem("##JumptoLocation", AppIcons.SIC_FileJumpTo, "Jump to file Location", [Colors.IconPrimary, Colors.IconPrimary, Colors.IconSecondary])) {
-            CurrentDir = bm.Path;
-        }
-        if (ImguiHelpers.ContextMenuItem("##RemoveBookmarks", AppIcons.SIC_BookmarkRemove, "Remove from Bookmarks", [Colors.IconPrimary, Colors.IconTertiary])) {
-            manager.RemoveBookmark(Workspace.Config.Game.name, bm.Path);
-        }
-        if (ImGui.BeginMenu($"{AppIcons.SI_GenericTag} | Tags")) {
-            ImGui.PushItemFlag(ImGuiItemFlags.AutoClosePopups, false);
-            foreach (var tag in BookmarkManager.TagInfoMap.Keys) {
-                bool hasTag = bm.Tags.Contains(tag);
-                if (ImGui.MenuItem(tag, "", hasTag)) {
-                    if (hasTag) {
-                        bm.Tags.Remove(tag);
-                        manager.SaveBookmarks();
-                    } else {
-                        bm.Tags.Add(tag);
-                        manager.SaveBookmarks();
-                    }
-                }
-            }
-            ImGui.Separator();
-            if (ImGui.MenuItem(Lang.Buttons.Clear_Tags)) {
-                bm.Tags.Clear();
-                manager.SaveBookmarks();
-            }
-            ImGui.PopItemFlag();
-            ImGui.EndMenu();
-        }
-        if (editingCustomBookmark != bm.Path) {
-            editingCustomBookmark = bm.Path;
-            customBookmarkComment = bm.Comment;
-        }
-        if (ImGui.InputTextWithHint("Edit Comment"u8, "Press Enter to save"u8, ref customBookmarkComment, 64, ImGuiInputTextFlags.EnterReturnsTrue)) {
-            bm.Comment = customBookmarkComment;
-            manager.SaveBookmarks();
-            editingCustomBookmark = null;
-        }
-    }
-
     private bool IsFileOrFolderInBundle(string path)
     {
         if (contentWorkspace.CurrentBundle?.HasResources != true) return false;
@@ -1156,14 +888,17 @@ public partial class PakBrowser(ContentWorkspace contentWorkspace, string[]? pak
 
         return parts[^1];
     }
-
-    private static string LessCompactFilePath(string path)
+    private void ToggleBookmarks()
     {
-        var parts = path.NormalizeFilepath().Split('/');
-        if (parts.Length <= 2) return path;
-
-        var remainingParts = parts[2..];
-        return string.Join("/", remainingParts);
+        if (AppConfig.Instance.UseBookmarkWindow) {
+            if (EditorWindow.CurrentWindow?.HasSubwindow<BookmarksWindow>(out var data) == true) {
+                EditorWindow.CurrentWindow.CloseSubwindow(data!);
+            } else {
+                EditorWindow.CurrentWindow?.AddUniqueSubwindow(new BookmarksWindow(BookmarksPanel));
+            }
+        } else {
+            isShowBookmarks = !isShowBookmarks;
+        }
     }
     public bool RequestClose()
     {
